+++ src/Lykke.Service.ChainalysisProxy.Services/ExposureDetails.cs
@@ -15,9 +15,9 @@ namespace Lykke.Service.ChainalysisProxy.Services
         }
 
         public string Category { get; set; }
-        public int SentIndirectExposure { get; set; }
-        public int SentDirectExposure { get; set; }
-        public int ReceivedIndirectExposure { get; set; }
-        public int ReceivedDirectExposure { get; set; }
+        public long SentIndirectExposure { get; set; }
+        public long SentDirectExposure { get; set; }
+        public long ReceivedIndirectExposure { get; set; }
+        public long ReceivedDirectExposure { get; set; }
     }
 }
